Move CollectionSelectedUseCase to account package; add UiDispatcher; use in ViewModels; update SYNC_DELAY to Duration; remove unused IoDispatcher from CollectionSelectedUseCase
