+++ core/src/main/kotlin/at/bitfire/davdroid/di/CoroutineDispatchersModule.kt
@@ -6,6 +6,7 @@ package at.bitfire.davdroid.di
 
 import at.bitfire.davdroid.di.qualifier.DefaultDispatcher
 import at.bitfire.davdroid.di.qualifier.IoDispatcher
+import at.bitfire.davdroid.di.qualifier.UiDispatcher
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -24,5 +25,9 @@ class CoroutineDispatchersModule {
     @Provides
     @IoDispatcher
     fun ioDispatcher(): CoroutineDispatcher = Dispatchers.IO
+
+    @Provides
+    @UiDispatcher
+    fun uiDispatcher(): CoroutineDispatcher = Dispatchers.Main.immediate
 
 }
