+++ core/src/main/kotlin/at/bitfire/davdroid/ui/account/CreateAddressBookViewModel.kt
@@ -12,12 +12,14 @@ import at.bitfire.davdroid.accounts.AccountId
 import at.bitfire.davdroid.accounts.toAndroidAccount
 import at.bitfire.davdroid.db.HomeSet
 import at.bitfire.davdroid.di.qualifier.ApplicationScope
+import at.bitfire.davdroid.di.qualifier.UiDispatcher
 import at.bitfire.davdroid.repository.DavCollectionRepository
 import at.bitfire.davdroid.repository.DavHomeSetRepository
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
@@ -26,7 +28,8 @@ class CreateAddressBookViewModel @AssistedInject constructor(
     @Assisted val accountId: AccountId,
     @ApplicationScope private val applicationScope: CoroutineScope,
     private val collectionRepository: DavCollectionRepository,
-    homeSetRepository: DavHomeSetRepository
+    homeSetRepository: DavHomeSetRepository,
+    @UiDispatcher private val uiDispatcher: CoroutineDispatcher
 ): ViewModel() {
 
     @AssistedFactory
@@ -77,7 +80,7 @@ class CreateAddressBookViewModel @AssistedInject constructor(
         val homeSet = uiState.selectedHomeSet ?: return
         uiState = uiState.copy(isCreating = true)
 
-        applicationScope.launch {
+        applicationScope.launch(uiDispatcher) {
             uiState = try {
                 collectionRepository.createAddressBook(
                     account = accountId.toAndroidAccount(),
