+++ core/src/main/kotlin/at/bitfire/davdroid/ui/account/CreateCalendarViewModel.kt
@@ -12,6 +12,7 @@ import at.bitfire.davdroid.accounts.AccountId
 import at.bitfire.davdroid.accounts.toAndroidAccount
 import at.bitfire.davdroid.db.HomeSet
 import at.bitfire.davdroid.di.qualifier.ApplicationScope
+import at.bitfire.davdroid.di.qualifier.UiDispatcher
 import at.bitfire.davdroid.repository.DavCollectionRepository
 import at.bitfire.davdroid.repository.DavHomeSetRepository
 import at.bitfire.synctools.icalendar.Css3Color
@@ -19,6 +20,7 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
@@ -35,7 +37,8 @@ class CreateCalendarViewModel @AssistedInject constructor(
     @Assisted val accountId: AccountId,
     @ApplicationScope private val applicationScope: CoroutineScope,
     private val collectionRepository: DavCollectionRepository,
-    homeSetRepository: DavHomeSetRepository
+    homeSetRepository: DavHomeSetRepository,
+    @UiDispatcher private val uiDispatcher: CoroutineDispatcher
 ): ViewModel() {
 
     @AssistedFactory
@@ -132,7 +135,7 @@ class CreateCalendarViewModel @AssistedInject constructor(
         val homeSet = uiState.homeSet ?: return
         uiState = uiState.copy(isCreating = true)
 
-        applicationScope.launch {
+        applicationScope.launch(uiDispatcher) {
             uiState = try {
                 collectionRepository.createCalendar(
                     account = accountId.toAndroidAccount(),
